requirements.txt
Flask
flask-cors
python-dotenv
google-genai
duckduckgo_search
pydantic
gunicorn
